code/cloud/requirements.txt
Flask
paho-mqtt
gunicorn
requests
opencv-python
stripe
pytz
Flask-SQLAlchemy
psycopg2-binary
Flask-SocketIO
gevent
gevent-websocket